Schritt 1
Conflicts:

	sw/source/core/doc/docedt.cxx
	sw/source/core/doc/doclay.cxx
	sw/source/core/doc/docnew.cxx
	sw/source/core/docnode/ndcopy.cxx
	sw/source/core/frmedt/fecopy.cxx
	sw/source/core/layout/frmtool.cxx
	sw/source/core/uibase/inc/content.hxx
	sw/source/core/uibase/utlui/content.cxx
	sw/source/core/uibase/wrtsh/wrtsh1.cxx
	sw/source/core/undo/rolbck.cxx
	sw/source/core/unocore/unochart.cxx
	sw/source/core/unocore/unocoll.cxx

+++ sw/source/core/undo/unattr.cxx
@@ -236,7 +236,7 @@ bool SwUndoFmtAttr::IsFmtInDoc( SwDoc* pDoc )
     const SwFmtsBase *fmts = NULL;
     bool bContains = false;
     sal_Int32 nPos;
-    sal_Int32 nMax = USHRT_MAX;
+    sal_Int32 nMax = UINT_MAX;
     switch ( m_nFmtWhich )
     {
         case RES_TXTFMTCOLL:
